+++ .eslintrc.js
@@ -68,6 +68,7 @@ module.exports = {
             'jest/require-to-throw-message': 'off',
             'jest/prefer-called-with': 'off',
             'jest/prefer-spy-on': 'off',
+            'jest/no-if': 'off',
         }
     }],
     settings: {