+++ src/index.ts
@@ -1,9 +1,9 @@
 import unfetch from 'isomorphic-unfetch'
-import { FetchError } from 'named-app-errors'
 
 import type { SerializedValue } from '@ergodark/types'
 
 export type FetchConfig = Omit<RequestInit, 'body'> & {
+    swr?: boolean,
     rejects?: boolean,
     ignoreParseErrors?: boolean,
     body?: Record<string, unknown>
@@ -24,7 +24,6 @@ let globalFetchConfig: FetchConfig = {
 /**
  * Re-export these
  */
-export { FetchError };
 export { unfetch };
 
 /**
@@ -42,53 +41,84 @@ export function setGlobalFetchConfig(config: FetchConfig) {
 }
 
 /**
- * Performs an isomorphic (un)fetch. Throws 1) when parsing the body for JSON
- * content fails and `config != { ignoreParseErrors: true }` or 2) when `config
- * = { rejects: true }` and a non-2xx response is received.
+ * Performs an isomorphic (un)fetch.
  *
- * Returns an HTTP Response object `res` and parsed response body `json`.
+ * Throws 1) when parsing the body for JSON content fails and `config != {
+ * ignoreParseErrors: true }` or 2) when `config = { rejects: true }` and a
+ * non-2xx response is received. Otherwise, returns an HTTP Response object
+ * `res` and parsed response body `json` on 2xx response or `error`.
+ *
+ * Example:
+ * ```
+ *   const { json, error } = fetch<{ myData: number }, { message: string }>('api/endpoint', {
+ *     method: 'POST',
+ *     body: b
+ *   });
+ *
+ *   ...
+ *
+ *   if(error) return error.message;
+ *   return json.myData + SOME_CONST;
+ * ```
  */
-export async function fetch<JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) {
-    const parsedOptions: RequestInit = {
+export async function fetch<
+  JsonType extends SerializedValue = Record<string, unknown>,
+  ErrorType extends SerializedValue = JsonType>(
+    url: RequestInfo,
+    config?: FetchConfig
+) {
+    const parsedOptions = {
         ...getGlobalFetchConfig(),
+        ...(config?.swr ? { method: 'GET' } : {}),
         ...config,
-        body: config?.body ? JSON.stringify(config.body) : undefined
+        body: config?.body !== undefined ? JSON.stringify(config.body) : undefined
     };
 
     const res = await unfetch(url, parsedOptions);
-    let json: JsonType | null = null;
+    let json: JsonType | undefined = undefined;
+    let error: ErrorType | undefined = undefined;
 
-    try { json = (await res.json()) || {} }
-    catch(err) { if(!config?.ignoreParseErrors) throw err; }
+    try { json = await res.json(); }
+    catch(err) { if(!parsedOptions?.ignoreParseErrors) throw err; }
 
-    if(!res.ok && config?.rejects)
-        throw new FetchError(res);
+    if(!res.ok) {
+        if(parsedOptions?.swr || parsedOptions?.rejects)
+            throw json;
 
-    return { res, json };
+        error = json as ErrorType;
+        json = undefined;
+    }
+
+    return { res, json, error };
 }
 
 /**
  * Syntactic sugar for calling `fetch(..., { method: 'GET', ... })`.
  */
-fetch.get = <JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) => {
-    return fetch<JsonType>(url, { method: 'GET', ...config });
-};
+fetch.get = <JsonType extends SerializedValue = Record<string, unknown>, ErrorType extends SerializedValue = JsonType>(
+  url: RequestInfo,
+  config?: FetchConfig
+) => fetch<JsonType, ErrorType>(url, { method: 'GET', ...config });
 
 /**
  * Syntactic sugar for calling `fetch(..., { method: 'PUT', ... })`.
  */
-fetch.put = <JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) => {
-    return fetch<JsonType>(url, { method: 'PUT', ...config });
-};
+fetch.put = <JsonType extends SerializedValue = Record<string, unknown>, ErrorType extends SerializedValue = JsonType>(
+  url: RequestInfo,
+  config?: FetchConfig
+) => fetch<JsonType, ErrorType>(url, { method: 'PUT', ...config });
 
 /**
  * Syntactic sugar for calling `fetch(..., { method: 'DELETE', ... })`.
  */
-fetch.delete = <JsonType extends SerializedValue>(url: RequestInfo, config?: FetchConfig) => {
-    return fetch<JsonType>(url, { method: 'DELETE', ...config });
-};
+fetch.delete = <JsonType extends SerializedValue = Record<string, unknown>, ErrorType extends SerializedValue = JsonType>(
+  url: RequestInfo,
+  config?: FetchConfig
+) => fetch<JsonType, ErrorType>(url, { method: 'DELETE', ...config });
 
 /**
  * Syntactic sugar for calling `fetch(..., { method: 'POST', ... })`.
  */
 fetch.post = fetch;
+
+fetch.swr = async (key: string) => (await fetch(key, { swr: true })).json;