+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 [Upcoming]: https://github.com/Xunnamius/isomorphic-json-fetch/compare/main...develop
+[v1.0.10]: https://github.com/Xunnamius/isomorphic-json-fetch/compare/v1.0.9...v1.0.10
 [v1.0.9]: https://github.com/Xunnamius/isomorphic-json-fetch/compare/v1.0.8...v1.0.9
 [v1.0.8]: https://github.com/Xunnamius/isomorphic-json-fetch/compare/v1.0.7...v1.0.8
 [v1.0.7]: https://github.com/Xunnamius/isomorphic-json-fetch/compare/v1.0.6...v1.0.7
@@ -31,6 +32,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 (no notes yet)
 
 ## Released
+
+### [![v1.0.10](https://api.ergodark.com/badges/github-tag-date/xunnamius/isomorphic-json-fetch/v1.0.10)][v1.0.10]
+#### Changed
+- Added better TypeScript support (ErrorType alongside JsonType)
+- Even better README.md with quick-start and explicit examples
+- SWR sugar
+- Updated documentation
+- More extensive testing
 
 ### [![v1.0.9](https://api.ergodark.com/badges/github-tag-date/xunnamius/isomorphic-json-fetch/v1.0.9)][v1.0.9]
 #### Changed